+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog adheres to semantic version according to [keepachangelog](https:/
 
 ## Unreleased
 
+### Added
+- Shorthand to `--body` flag: `-B`
+- Ability to read body from stdin
+
 ### Changed
 - Rename sub-command `url` to `alias`